refs #1242, CAP-90, CAP-94 - yass - Add more access-control test cases (e.g. various sequences of initialization/authorization/deauthorization)
Note that the test case for "authorized->deauthorized" is interesting. It works with LocalizedMemory replicas but not GenericSQL replicas; this is because LocalizedMemory consistently saves/returns tombstones without any extra programming, but GenericSQL must be enhanced to consistently support tombstones. The handling of tombstones should be addressed more thoroughly as part of the CAP-94.