refs #1242, CAP-123 - yass - Add support definint a replica's "effective" replica ID. Any entities updated or created will use the "effective" ID instead of the actual ID.
Note: This is require to make proxied replicas use the right IDs for purposes of join/rejoin/reset (and, to lesser extent, sync) operations. However, this is only necessary because of certain implementation decisions for those operations. If those operations were reimplemented, then this bit of hackery may become unnecessary.